+++ modules/addonlib/src/main/scala/docspell/addons/out/NewItem.scala
@@ -81,7 +81,8 @@ object NewItem {
         reprocess = false,
         attachmentsOnly = attachmentsOnly,
         customData = customData,
-        process = None
+        process = None,
+        runAddons = None
       )
   }