+++ modules/joex/src/main/scala/docspell/joex/scanmailbox/ScanMailboxTask.scala
@@ -331,6 +331,7 @@ object ScanMailboxTask {
           None,
           None,
           None,
+          None,
           None
         )
         data = OUpload.UploadData(